+++ assets/src/block-editor/blocks/hand-picked-posts/edit.js
@@ -27,7 +27,7 @@ import PostsPicker from './components/posts-picker';
 import HandpickedPostBlockControls from './components/block-controls';
 import InspectorControls from '../common-posts-list/components/inspector-controls';
 import './editor.css';
-import EditWithSelectMultiplePostTypes from '../common-posts-list/edit-with-select-multiple-post-types';
+import { EditWithApiFetch } from '../common-posts-list/edit';
 import getConfig from '../../utils/get-config';
 
 /**
@@ -60,7 +60,7 @@ const Edit = props => {
 			{ editMode ? (
 				<PostsPicker { ...props } />
 			) : (
-				<EditWithSelectMultiplePostTypes { ...props } />
+				<EditWithApiFetch { ...props } />
 			) }
 		</>
 	);
